+++ src/compile.ts
@@ -24,6 +24,7 @@ export type {
   CompilationResult,
   ExecutableTarget,
   ComplexResult,
+  CompiledValue,
   CompiledRunner,
   ExpressionRunner,
   LambdaRunner,